Moved RunSample method to the top of the file for easy access.
