+++ Samples/ResourceManagement/Cdn/ManageCdn/Program.cs
@@ -13,7 +13,7 @@ using System;
 using System.Collections.Generic;
 using System.Net.Http;
 
-namespace ManageRedis
+namespace ManageCdn
 {
     public class Program
     {
